Some Fixes are done to improve the monitoring

+++ child_app/android/app/src/main/kotlin/com/example/child_app/BootReceiver.kt
@@ -23,7 +23,7 @@ class BootReceiver : BroadcastReceiver() {
             )
             
             // Flutter shared_preferences "flutter." prefix lagate hain
-            val childId = prefs.getInt("flutter.child_id", -1)
+            val childId = prefs.getLong("flutter.child_id", -1L).toInt()
 
             Log.d("BOOT_RECEIVER", "Child ID from prefs: $childId")
